Updated bitwarden_passwordless javascript function
Added the async helper function `register` that performs the
frontend registration process of the user's passkey.

Removed comparing the returned token against the previous token
to determine when a value should be returned to Streamlit since
the function will only be called when a button is clicked in Streamlit.
Added error handling and console logging.