Updated func bitwarden_passwordless.frontend._register
Removed the return value since the return value from the javascript function
is saved to the session state with a key defined by the `key` parameter.
The return value from the function is always None for some reason and the
only way to access the true return value is by retrieving it from the session state.

Also updated `bitwarden_passwordless.client.BitwardenPasswordlessClient.register-user`
in a similar way since it uses the `_register` function.