Updated BitwardenRegisterConfig.expires_at
Changed the parameter `expires_at` to be a `timedelta` rather than a
`datetime` object, which allows the user to specify for how long the
registration token will be valid from the start of the registration process.
Renamed the parameter to `validity`. Since the register config is initialized
when the `BitwardenPasswordlessClient` is initialized a timedelta is necessary
because we don't know when a user will trigger the registration process.

Added the property `expires_at`, which defines the expiry datetime of the
registration token based on the `validity` parameter. Updated and added
unit tests to reflect the changes.