Updated javascript function to v0.3.0
Redesigned the function to render a register/sign in button that gets
assigned an onClick callback function to either start the register
(action='register') or sign in (action='sign_in') process. By clicking a
button rendered on the javascript side instead of the Streamlit side the
"Document is not focused" error that is raised when trying to register or
sign in from an iPhone will hopefully be resolved. Also added a parameter
to be able to disable the register/sign in button.