First re-designed stage of bitwarden_register_from
Removed the Streamlit form and replaced the form submit button
with the `streamlit_passwordless.register_button` to trigger the
register process. The helper function `_validate_form` has been
replaced by the functions `_validate_username` and `_create_user`.
A solution to handle if the register_button was clicked or not
is needed.