Added streamlit form to register_from component
The streamlit form allows the user to enter all user form fields before
rerunning the application, which ensures a new user is only created once
when all desired fields are filled in. Before a rerun was triggered whenever
a form field was filled in causing a new user to be created each time. This
caused a mismatch in the user_id registered with Bitwarden Passwordless and
the user_id saved to the database.

A form validation button has been added and when the validation passes the
register button is enabled.