Added the _create_register_token helper function to register_form
A cached function for creating a register token to register a new passkey
with the user's device. When used avoids calling the create_register_token
endpoint of the Bitwarden Passwordless API again between the click
of the `validate` and the `register_button`.

Added hash methods to `bitwarden_passwordless.BitwardenPasswordlessClient`
and `models.User` to allow them to be hashed in the `st.cache_data` decorator.