Fixing README.rst to pass PyPI rst validation
Removed the `:scale:` parameter of the badge images, since using
`:scale:` makes the PyPI reStructuredText validation fail even though
the `twine check` validation passes when `:scale:` is present.
Using `:scale:` passed the PyPI validation for the 0.6.0 release of
streamlit-passwordless on 2024-08-04.

It makes no difference to remove `:scale:` since it was set to 100% anyway.