Added discoverability widget to bitwarden_register_form
The discoverability widget is a toggle switch or radio button, which allows the
user to toggle between creating a discoverable or non-discoverable passkey.
If the discoverability_widget is added to the form the value from the widget
is passed on to creation of the register token and overrides the default
value for discoverability set for `BitwardenPasswordlessClient.register_config`.