Updated default help text handling of bitwarden_register_form
Moved the default help texts for the `username`, `displayname` and `alias` fields
to constants just like the fields for `credential_nickname` and `discoverability`.

Updated the username default help text to not suggest the username being an email,
because we should not encourage users to give away their email addresses without
a reason. Changed the username placeholder to not be an email address.