Fixed failing unit test for bitwarden_passwordless.frontend.sign_in_button
The session state was modified during the test for `register_button`
and carried over to the test for the `sign_in_button`.

Added fixture `mocked_session_state` which mocks the streamlit
session state with an empty dictionary to properly isolate the tests.
Updated tests for `sign_in_button` and `register_button` to use this
fixture and also to test that the buttons can be identified as clicked
or not clicked correctly.