Added setting database schema through env var STP_DB_SCHEMA
By specifying the environment variable `STP_DB_SCHEMA` the schema
to use for the tables of the database can be specified. If not specified
no schema is used. When using `__table_args__` for constraints e.g.
`UniqueConstraint` we cannot set schema in `__table_args__` of the `Base`
model because it may be overridden by subclasses.