Updated helper func _create_user of bitwarden_register_form
Removed the `user_id` parameter since creating a new user will always
create a new user_id. If a user already exists in the database the
`streamlit_passwordless.User` should be created from the database user
model instead.

Added the optional `role` parameter to assign a role to the new user.
If not specified the default role of `streamlit_passwordless.User`
is used.

Removed capturing `StreamlitPasswordlessError` and returning None. This
exception should not occur and it is better to let the application
crash loudly if it gets raised. Another benefit is that a "None check"
on the return value is no longer needed.