Fixed handling state in bitwarden passwordless custom component
The bitwarden passwordless custom Streamlit component would get
re-initialized when another widget on a page got clicked. This resulted
in a fresh reload of all JavaScript defined in the component
and the global state variables would go back to their default values.
This is potentially a bug which is also described in issue
https://github.com/streamlit/streamlit/issues/2145, but does not seem to
have been resolved.

If the sign-in/register button was clicked or not was determined by the
global variable `buttonNrClicks` and when the JavaScript got reloaded
it was reset to 0 and the button was not recognized as clicked.

Updated the custom component to compare the returned sign-in/register
token  against the token from the previous run to determine if the button
was clicked or not. This eliminates keeping the state of the component
on the JavaScript side and resolves the "clicked" issue.