Updated func database.cache.init
Removed parameter `if_exists` and the possibility to re-raise
`streamlit_passwordless.DatabaseError` if the user roles already exist
in the database. If an exception is raised there will be no return value
to cache and the `st.cache_resource` decorator will not work as expected.

Changed the return value from None to the following:

 - error : bool
    True if a `streamlit_passwordless.DatabaseError` occurred and the
    database could not be initialized correctly and False for no error.

- error_msg : str
    An error message that is safe to display to the user. An empty
    string is returned if `error` is False.