Added authorization to app.pages.admin
Added func `admin_page_authorized`, which is executed when the page module
is run as script or Streamlit page. It is defined as `admin_page` decorated
with `authorization.authorized` to only allows admin users to access its
contents. If not authorized the user is redirected to the sign in page.
`admin_page_authorized` is used by the CLI command `stp run admin`. The
func `admin_page` is unchanged to let the user of `streamlit_passwordless`
use `admin_page` in his application and apply the `authorization.authorized`
decorator as desired.