Added bitwarden_sign_in_button component
Allows the user to sign in to the application with a discoverable passkey.
Similar to :func:`bitwarden_sign_in_form`, but without the possibility to
sign in with user_id, alias or a non-discoverable passkey. It also provides
the `redirect` parameter, which will redirect the user to specified page on
successful sign in. Moved the sign in logic and authentication logic to the
helper func `_process_user_sign_in`.