Updated bitwarden_sign_in_form with functionality from bitwarden_sign_in_button
Added the `banner_container`, parameter, which allows the user to pass
in a custom container in which the error and success messages from the
sign in form will be written to. Added the functionality to redirect user
to specified page on successful sign in. Applied the new sign in logic and
authentication from helper func `_process_user_sign_in`.