Updated func database.crud.user.get_user_by_username
Changed the `disabled` parameter to accept bool or None.
When None filtering by disabled or enabled user is omitted.
This is important when you want to check if a user exists regardless
of the state being disabled or enabled, which is the case for the
register and create user forms.