Updated user email fixtures
Changed `email_id` attribute to None in fixtures `user_1_email_primary`
and `user_1_email_secondary`, which will model the state of the emails
before being saved to the database (that case will be tested). Removed
the UTC timezone info of the timestamp columns since this information
will not be present when an email is loaded from the database.