Added default value of '' to models.Email.user_id
By using '' as a default value of `models.Email.user_id` it makes
it easier to create a new `models.User` with an email address, since the
`user_id`` will be generated when `models.User` is initialized and will not
be known in advance. It is not necessary to specify `models.Email.user_id`
when saving an email to the database since the user_id will be derived
from its user.