Updated database.crud.user.create_user
Added parameter `custom_roles`, which accepts a sequence of
`database.models.CustomRole` that will take precedence over the
custom roles defined on `user`. This avoids a database lookup for
the custom roles to associate with `user` since they already exist
in the current database session.

Updated the related function `components.core.create_user_in_database`
and added a unit test for this.