Implemented saving custom roles to db in create_user_form
Changed the parameters related to custom roles to accept
`database.models.CustomRole` instead of `models.CustomRole`, which makes
it easier to associate the user with the custom roles in the database.
If `custom_roles` parameter is not specified the defined custom roles
in the database will be loaded and made available for selection.