Improved handling of and added new default table columns
Added the mixin class `database.models.ModifiedAndCreatedColumnMixin`,
which adds the columns `modified_at`, `modified_by`, `created_at` and
`created_by` to a database model. Applied the mixin to the models `Role`,
`CustomRole`, `User` and `Email`. Removed column `modified_at` from
`UserSignIn` because records in this table are created on user sign in
and should not be modified after creation. Updated unit tests.