Added param require_credential_nickname to bitwarden register forms
It is easy to enter a credential nickname and forget to press enter
to apply the value, meaning that a passkey without a nickname will be
registered. Setting `require_credential_nickname` to True (the default)
disables the register button if the credential_nickname field is not set.
Added parameter `require_credential_nickname` to `bitwarden_register_form`
and `bitwarden_register_form_existing_user`.