Updated func database.core.create_session_factory
Removed returning the database engine as a second argument, since the
argument is almost always ignored when using the function. Added support
to pass a custom declarative base class to use when creating the database.
This is done through the `base` parameter. Added the parameter `connect_args`
that will be passed along to the `connect_args` parameter of `sqlalchemy.create_engine`.