Changed dtype of db column user.user_id from str to UUID
Having a proper UUID datatype for the user.user_id column allows databases
(that support UUID) to leverage this datatype for improved performance and
reduced storage space. `Mapped[uuid.UUID]` maps to `sqlalchemy.Uuid`,
which can be represented as TEXT or VARCHAR for databases that do not support
UUID datatypes.

Updated the pydantic models to use `uuid.UUID` as their datatype for the
`user_id` columns. Updated unit tests and type annotations.