Updated development environment
Changed to use `ruff` as the formatter and linter of the project, which
replaces `black`, `flake8` and `isort`. Added ruff linting rules.
Newfound linting errors will be fixed progressively when updates occur.
Updated `mypy` rules.

Updated the development dependencies and the conda dev environment.
Changed the format of the dev environment lock file from `environment.yml`
to the `explicit` format, which further locks down the exact channel that
is used for each package.

Updated to require `setuptools` >= v77.0 and `twine` >=6.2 to support the
new SPDX license identifier syntax in `pyproject.toml`.