Added func database.models.audit_columns_mixin_factory
Creates the mixin class for the audit columns.
The audit columns consists of `updated_at`, `updated_by`, `created_at`
and `created_by`. It is used to customize the behavior of (un)deferring
loading the audit columns by default. Deferred is the new default state of
these columns. `ModifiedAndCreatedColumnMixin`, which has been renamed to
`AuditColumnsMixin`, is created from the function.

Added unit tests for testing `AuditColumnsMixin` and removed testing
the audit columns in other tests.