Updated config of Base database model
Set mapper argument `eager_defaults` to True to enable returning new
values for the audit columns `updated_at` and `created_at` in the same
flush operation as an update or insert instead of fetching these values
on subsequent access. These columns will be in deferred state when lazy
loaded after after an update or insert even though they were originally
loaded as undeferred. This change better reflects the expected behavior
of working with the audit columns.